Refactor audio feature and home navigation
