+++ lib/core/di/injection.dart
@@ -1,11 +1,11 @@
 
 import 'package:get_it/get_it.dart';
-import '../../data/repositories/audio_repository.dart';
 import '../services/audio_service.dart';
 import '../services/storage_service.dart';
 import '../../features/adhkar_reminder/data/services/adhkar_reminder_service.dart';
 import '../../features/adhkar_reminder/presentation/cubit/adhkar_reminder_cubit.dart';
-import '../../features/audio/bloc/audio_bloc.dart';
+import '../../features/audio/domain/services/audio_favorites_service.dart';
+import '../../features/audio/domain/services/sleep_timer_service.dart';
 import '../../features/prayer_times/data/repositories/prayer_times_repository.dart';
 import '../../features/prayer_times/data/services/prayer_notification_service.dart';
 import '../../features/prayer_times/presentation/cubit/prayer_times_cubit.dart';
@@ -21,6 +21,9 @@ Future<void> setupDependencies() async {
   
   // Services
   getIt.registerLazySingleton<AudioService>(() => AudioService());
+  getIt.registerLazySingleton<AudioFavoritesService>(
+      () => AudioFavoritesService());
+  getIt.registerLazySingleton<SleepTimerService>(() => SleepTimerService());
   getIt.registerLazySingleton<AdhkarReminderService>(
       () => AdhkarReminderService());
   getIt.registerLazySingleton<PrayerNotificationService>(
@@ -28,15 +31,10 @@ Future<void> setupDependencies() async {
   getIt.registerLazySingleton<WallpapersService>(() => WallpapersService());
 
   // Repositories
-  getIt.registerLazySingleton<AudioRepository>(() => AudioRepositoryImpl());
   getIt.registerLazySingleton<PrayerTimesRepository>(
       () => PrayerTimesRepository());
 
   // BLoCs
-  getIt.registerFactory<AudioBloc>(() => AudioBloc(
-    audioRepository: getIt<AudioRepository>(),
-    audioService: getIt<AudioService>(),
-  ));
   getIt.registerFactory<TasbihCubit>(() => TasbihCubit());
   getIt.registerFactory<AdhkarReminderCubit>(
       () => AdhkarReminderCubit(service: getIt<AdhkarReminderService>()));
