+++ lib/features/audio/domain/services/sleep_timer_service.dart
@@ -1,0 +1,35 @@
+import 'dart:async';
+
+typedef SleepTimerCallback = void Function(int minutes);
+
+class SleepTimerService {
+  Timer? _timer;
+  int? _minutes;
+
+  bool get isActive => _timer != null;
+  int? get minutes => _minutes;
+
+  void startTimer({
+    required int minutes,
+    required SleepTimerCallback onTimerComplete,
+  }) {
+    cancelTimer();
+    _minutes = minutes;
+    _timer = Timer(Duration(minutes: minutes), () {
+      final completedMinutes = _minutes ?? minutes;
+      _timer = null;
+      _minutes = null;
+      onTimerComplete(completedMinutes);
+    });
+  }
+
+  void cancelTimer() {
+    _timer?.cancel();
+    _timer = null;
+    _minutes = null;
+  }
+
+  void dispose() {
+    cancelTimer();
+  }
+}
