--- lib/features/audio/widgets/search_bar_widget.dart
@@ -1,85 +1,0 @@
-
-import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import '../../../core/utils/app_colors.dart';
-
-class SearchBarWidget extends StatefulWidget {
-  final String categoryName;
-  final Function(String) onSearch;
-
-  const SearchBarWidget({
-    Key? key,
-    required this.categoryName,
-    required this.onSearch,
-  }) : super(key: key);
-
-  @override
-  State<SearchBarWidget> createState() => _SearchBarWidgetState();
-}
-
-class _SearchBarWidgetState extends State<SearchBarWidget> {
-  final TextEditingController _searchController = TextEditingController();
-  bool _isSearching = false;
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(25),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            blurRadius: 8,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: TextField(
-        controller: _searchController,
-        textDirection: TextDirection.rtl,
-        textAlign: TextAlign.right,
-        decoration: InputDecoration(
-          hintText: 'البحث في ${widget.categoryName}...',
-          hintStyle: TextStyle(
-            color: Colors.grey[500],
-            fontSize: 16,
-          ),
-          prefixIcon: _isSearching
-              ? IconButton(
-                  onPressed: _clearSearch,
-                  icon: const Icon(Icons.clear, color: Colors.grey),
-                )
-              : const Icon(FontAwesomeIcons.search, color: Colors.grey),
-          border: InputBorder.none,
-          contentPadding: const EdgeInsets.symmetric(
-            horizontal: 20,
-            vertical: 16,
-          ),
-        ),
-        onChanged: (value) {
-          setState(() {
-            _isSearching = value.isNotEmpty;
-          });
-          widget.onSearch(value);
-        },
-        onSubmitted: widget.onSearch,
-      ),
-    );
-  }
-
-  void _clearSearch() {
-    _searchController.clear();
-    setState(() {
-      _isSearching = false;
-    });
-    widget.onSearch('');
-  }
-}
